Actualización a versión más actual
git-svn-id: https://svn.code.sf.net/p/hotels-game/code@5 cbd89f15-8f70-4dcd-96d5-802d9ad3bda2

+++ Dado.cs
@@ -7,17 +7,17 @@ namespace Juego_Hotel
 {
     public class Dado
     {
-        short n_caras;
+        int n_caras;
         Random rand;
-        public Dado(short caras)
+        public Dado(int caras)
         {
             this.n_caras = caras;
             this.rand = new Random();
         }
 
-        public short tirar()
+        public int tirar()
         {
-            return (short) this.rand.Next(1, this.n_caras+1);
+            return this.rand.Next(1, this.n_caras+1);
         }
     }
 }
